backend/requirements.txt
# PROVENIQ Properties - Backend Dependencies
# Python 3.11+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Auth
firebase-admin==6.4.0
python-jose[cryptography]==3.3.0

# Storage (provider interface)
google-cloud-storage==2.14.0
boto3==1.34.14

# PDF Generation
reportlab==4.0.8
weasyprint==60.2

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
